countries data fetched. india data fetched. state data fetched.

+++ app/src/main/java/com/example/covidtracker/Models/CountryInfo.kt
@@ -5,6 +5,6 @@ data class CountryInfo(
     val flag: String,
     val iso2: String,
     val iso3: String,
-    val lat: Int,
-    val long: Int
+    val lat: Float,
+    val long: Float
 )
